auth_service/requirements.txt
fastapi~=0.119.1
SQLAlchemy~=2.0.44
PyJWT~=2.10.1
python-dotenv~=1.1.1
starlette~=0.48.0
sqlmodel~=0.0.27
pydantic~=2.12.3
requests~=2.32.5
bcrypt~=5.0.0
dnspython~=2.8.0
idna~=3.11
anyio~=4.11.0
h11~=0.16.0
pip~=23.2.1
attrs~=25.4.0
Jinja2~=3.1.6
pytz~=2025.2
typing_extensions~=4.15.0
sniffio~=1.3.1
MarkupSafe~=3.0.3
Werkzeug~=3.1.3
click~=8.1.8
blinker~=1.9.0
itsdangerous~=2.2.0
certifi~=2025.10.5
pydantic_core~=2.41.4
email-validator~=2.3.0
urllib3~=2.5.0
uvicorn~=0.38.0
greenlet~=3.2.4
psycopg2-binary~=2.9.11
annotated-types~=0.7.0
charset-normalizer~=3.4.4
aniso8601~=10.0.1
Flask~=3.1.2
jsonschema~=4.25.1
referencing~=0.37.0
jsonschema-specifications~=2025.9.1
flask-restplus~=0.13.0
six~=1.17.0

dotenv~=0.9.9
pika~=1.3.2